final moving of isDevChain()
