+++ libraries/psibase/common/include/psibase/Rpc.hpp
@@ -78,6 +78,56 @@ namespace psibase
 
    };
 
+   /// Helper function to check if a URL contains localhost
+   inline bool isFromLocalhost(const std::string& url)
+   {
+      std::string_view value = url;
+      
+      // Trim whitespace
+      auto leading = value.find_first_not_of(" \t");
+      auto trailing = value.find_last_not_of(" \t");
+      if (leading != std::string_view::npos)
+      {
+         value = value.substr(leading, trailing - leading + 1);
+      }
+      else
+      {
+         return false;
+      }
+
+      // Find position of scheme separator
+      size_t pos = value.find("://");
+      std::string_view domain;
+      
+      if (pos != std::string_view::npos)
+      {
+         // Create a string view starting after "://"
+         domain = value.substr(pos + 3);
+      }
+      else
+      {
+         // No scheme found, treat entire value as domain
+         domain = value;
+      }
+
+      // Find the end of the domain name (either at first colon or slash, or end of string)
+      size_t domainEndPos = domain.find(':');
+      if (domainEndPos == std::string_view::npos)
+      {
+         domainEndPos = domain.find('/');
+      }
+      if (domainEndPos != std::string_view::npos)
+      {
+         domain = domain.substr(0, domainEndPos);
+      }
+
+      return domain.find("localhost") != std::string_view::npos;
+   }
+
+   /// Checks if the request origin indicates a development chain
+   /// by looking for "localhost" in the origin header domain
+   bool isDevChain(const HttpRequest& request);
+
    struct URIPath
    {
       std::string path;
