+++ services/system/HttpServer/src/HttpServer.cpp
@@ -14,92 +14,6 @@ static constexpr bool enable_print = false;
 
 using namespace psibase;
 
-namespace
-{
-   // Used to convert a split_view element to a string_view.
-   // Note that this is only needed when the standard library
-   // doesn't implement P2210R2
-   std::string_view split2sv(const auto& r)
-   {
-      auto data = &*r.begin();
-      auto size = static_cast<std::size_t>(std::ranges::distance(r));
-      return std::string_view{data, size};
-   }
-}  // namespace
-
-namespace psibase
-{
-   bool isLocalhost(const std::string& url)
-   {
-      std::string_view value = url;
-      
-      // Trim whitespace
-      auto leading = value.find_first_not_of(" \t");
-      auto trailing = value.find_last_not_of(" \t");
-      if (leading != std::string_view::npos)
-      {
-         value = value.substr(leading, trailing - leading + 1);
-      }
-      else
-      {
-         return false;
-      }
-
-      // Find position of scheme separator
-      size_t pos = value.find("://");
-      std::string_view domain;
-      
-      if (pos != std::string_view::npos)
-      {
-         // Create a string view starting after "://"
-         domain = value.substr(pos + 3);
-      }
-      else
-      {
-         // No scheme found, treat entire value as domain
-         domain = value;
-      }
-
-      // Find the end of the domain name (either at first colon or slash, or end of string)
-      size_t domainEndPos = domain.find(':');
-      if (domainEndPos == std::string_view::npos)
-      {
-         domainEndPos = domain.find('/');
-      }
-      if (domainEndPos != std::string_view::npos)
-      {
-         domain = domain.substr(0, domainEndPos);
-      }
-
-      return domain.find("localhost") != std::string_view::npos;
-   }
-
-   bool isDevChain(const HttpRequest& request)
-   {
-      // Check the host field first (always present)
-      if (isLocalhost(request.host))
-      {
-         return true;
-      }
-      
-      // Also check origin header if present (for cross-origin requests)
-      if (auto originHeader = request.getHeader("origin"))
-      {
-         for (auto part : *originHeader | std::views::split(','))
-         {
-            std::string_view value = split2sv(part);
-            std::string valueStr{value};
-            
-            if (isLocalhost(valueStr))
-            {
-               return true;
-            }
-         }
-      }
-      return false;
-   }
-}  // namespace psibase
-
 namespace SystemService
 {
    namespace
